+++ CHANGELOG.md
@@ -6,5 +6,6 @@ Unreleased
 
  - System query is now supported when the System registry is present.
  - Service query supports the `--show-service-uri/-u` flag.
+ - Services can be inspected to get more information. 
  - Setup and installation now set up the docker subnetwork and container ip addresses.
  - Setup, installation, and running will now default to installing local clouds in `~/.config/pyrrowhead/local-clouds`, but this can be overridden if another install location is desired.